CTA now has more or less columns depending on viewport size and some other minor changes as removing unused declarations
