+++ src/components/Layout.js
@@ -1,4 +1,4 @@
-import { Box, Button, useColorModeValue } from '@chakra-ui/react';
+import { Box } from '@chakra-ui/react';
 import Head from 'next/head';
 
 import Navbar from './navigation/Navbar';
