+++ src/components/hero/cta/Timeslots.js
@@ -1,0 +1,47 @@
+import { Box, Flex, SimpleGrid, Text, VStack } from '@chakra-ui/react';
+import { Link } from '@chakra-ui/next-js';
+import { FaBolt } from 'react-icons/fa';
+import { TbMoonFilled } from 'react-icons/tb';
+
+import Slot from './Slot';
+
+const Timeslots = () => {
+	return (
+		<VStack align={'start'}>
+			{/* Header */}
+			<Flex alignItems={'center'} direction={'row'} wrap={'wrap'} mb={'.9rem'}>
+				<Text as={'span'} fontWeight={'thin'}>
+					Schedule your collection in{' '}
+					<Text as={'span'} fontWeight={'bold'}>
+						London
+					</Text>
+				</Text>
+			</Flex>
+			{/* Content */}
+			<SimpleGrid
+				className="slots"
+				w={'100%'}
+				columns={{ base: 1, md: 2 }}
+				spacing={2}
+			>
+				<Slot
+					url={'#'}
+					id={'earliest'}
+					icon={FaBolt}
+					timeslot={'In the next 45min'}
+				/>
+				<Slot
+					url={'#'}
+					id={'last'}
+					icon={TbMoonFilled}
+					timeslot={'20:00 - 23:00'}
+				/>
+				<Box className="footer">
+					<Link href={'#'}>See all slots</Link>
+				</Box>
+			</SimpleGrid>
+		</VStack>
+	);
+};
+
+export default Timeslots;
